Cap max zoom.
Why?  Nonsensical to zoom in much more, and easy to get in that area.

Ideally this would cap at a max absolute value (e.g., 10 days).  With this approach, the abs value of max zoomed in depends on input data.  But, it seems good enough, max data range is fairly well known (700d, 1929).