Show label at biggest drop point.
Currently showing it at the last point, but the drop value it shows is for the max drop.  Current and new coincide given the current data, but that may change in the future, and has the exception of the latest (currentLabel / this year's) downturn, for which data continues below the min.

The date shown in the label is still the start / ATH date, so this isn't the date for the point the label is shown at either.  Maybe we want to change that too. Delaying that as I think we eventually we need to revamp the input data.  We want more of the original pandas dataframe and do some further manipulation in ts.